renamed main package to src
