+++ src/XAI/explainability.py
@@ -10,9 +10,9 @@ from pytorch_grad_cam.utils.image import show_cam_on_image, preprocess_image
 from pytorch_grad_cam.utils.model_targets import ClassifierOutputTarget
 from torch import Tensor
 from torch.utils.data import Subset, DataLoader
-from recaptcha_classifier import (
+from src import (
     DetectionLabels, DataPreprocessingPipeline, MainCNN)
-from recaptcha_classifier.XAI.utils_wrapped_model import WrappedModel
+from src.XAI.utils_wrapped_model import WrappedModel
 
 
 class Explainability(object):
