+++ src/models/main_model/model_class.py
@@ -1,7 +1,7 @@
 import torch
 import torch.nn as nn
-from recaptcha_classifier.constants import INPUT_SHAPE
-from recaptcha_classifier.models.base_model import BaseModel
+from src.constants import INPUT_SHAPE
+from src.models.base_model import BaseModel
 
 
 class MainCNN(BaseModel):
