+++ tests/unit/data/test_loader_factory.py
@@ -1,14 +1,14 @@
 import unittest
 from pathlib import Path
 from unittest.mock import patch
-from recaptcha_classifier.data.loader_factory import LoaderFactory
-from recaptcha_classifier.data.preprocessor import ImagePrep
-from recaptcha_classifier.data.augment import AugmentationPipeline
+from src.data.loader_factory import LoaderFactory
+from src.data.preprocessor import ImagePrep
+from src.data.augment import AugmentationPipeline
 from torch.utils.data import DataLoader
 
 
 class TestLoaderFactory(unittest.TestCase):
-    @patch("recaptcha_classifier.data.loader_factory.ImageDataset")
+    @patch("src.data.loader_factory.ImageDataset")
     def test_create_loaders(self, dataset_mock):
         class_map = {"class1": 0, "class2": 1}
         preprocessor = ImagePrep()
